refactor: extract types and constants to separate files per code review and reverted version changes
